Feat: delete namespace when uninstall argocd & kube-prometheus
