+++ internal/pkg/plugin/argocd/reinstall.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/merico-dev/stream/internal/pkg/log"
 	"github.com/merico-dev/stream/pkg/util/helm"
+	"github.com/merico-dev/stream/pkg/util/k8s"
 )
 
 func Reinstall(options *map[string]interface{}) (bool, error) {
@@ -32,6 +33,17 @@ func Reinstall(options *map[string]interface{}) (bool, error) {
 		return false, err
 	}
 
+	// delete the namespace
+	kubeClient, err := k8s.NewClient()
+	if err != nil {
+		return false, err
+	}
+	if err = kubeClient.DeleteNamespace(param.Chart.Namespace); err != nil {
+		log.Errorf("Failed to delete the %s namespace: %s", param.Chart.Namespace, err)
+		return false, err
+	}
+
+	// install
 	log.Info("Installing or updating argocd helm chart ...")
 	if err = h.InstallOrUpgradeChart(); err != nil {
 		return false, err
