+++ internal/pkg/plugin/argocd/uninstall.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/merico-dev/stream/internal/pkg/log"
 	"github.com/merico-dev/stream/pkg/util/helm"
+	"github.com/merico-dev/stream/pkg/util/k8s"
 )
 
 func Uninstall(options *map[string]interface{}) (bool, error) {
@@ -29,6 +30,16 @@ func Uninstall(options *map[string]interface{}) (bool, error) {
 
 	log.Info("uninstalling argocd helm chart")
 	if err = h.UninstallHelmChartRelease(); err != nil {
+		return false, err
+	}
+
+	// delete the namespace
+	kubeClient, err := k8s.NewClient()
+	if err != nil {
+		return false, err
+	}
+	if err = kubeClient.DeleteNamespace(param.Chart.Namespace); err != nil {
+		log.Errorf("Failed to delete the %s namespace: %s", param.Chart.Namespace, err)
 		return false, err
 	}
